+++ sdk/anomalydetector/ai-anomaly-detector-rest/src/index.ts
@@ -11,6 +11,5 @@ export * from "./isUnexpected";
 export * from "./models";
 export * from "./outputModels";
 export * from "./paginateHelper";
-export * from "./pollingHelper";
 
 export default AnomalyDetectorRest;