+++ sdk/digitaltwins/arm-digitaltwins/samples/v3/javascript/privateEndpointConnectionsListSample.js
@@ -15,7 +15,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to List private endpoint connection properties.
  *
  * @summary List private endpoint connection properties.
- * x-ms-original-file: specification/digitaltwins/resource-manager/Microsoft.DigitalTwins/stable/2022-05-31/examples/PrivateEndpointConnectionsList_example.json
+ * x-ms-original-file: specification/digitaltwins/resource-manager/Microsoft.DigitalTwins/stable/2022-10-31/examples/PrivateEndpointConnectionsList_example.json
  */
 async function listPrivateEndpointConnectionProperties() {
   const subscriptionId = "50016170-c839-41ba-a724-51e9df440b9e";