+++ sdk/postgresql/arm-postgresql/src/operations/serverSecurityAlertPolicies.ts
@@ -6,7 +6,8 @@
  * Changes may cause incorrect behavior and will be lost if the code is regenerated.
  */
 
-import { PagedAsyncIterableIterator } from "@azure/core-paging";
+import { PagedAsyncIterableIterator, PageSettings } from "@azure/core-paging";
+import { setContinuationToken } from "../pagingHelper";
 import { ServerSecurityAlertPolicies } from "../operationsInterfaces";
 import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers";
@@ -18,12 +19,12 @@ import {
   ServerSecurityAlertPolicy,
   ServerSecurityAlertPoliciesListByServerNextOptionalParams,
   ServerSecurityAlertPoliciesListByServerOptionalParams,
+  ServerSecurityAlertPoliciesListByServerResponse,
   SecurityAlertPolicyName,
   ServerSecurityAlertPoliciesGetOptionalParams,
   ServerSecurityAlertPoliciesGetResponse,
   ServerSecurityAlertPoliciesCreateOrUpdateOptionalParams,
   ServerSecurityAlertPoliciesCreateOrUpdateResponse,
-  ServerSecurityAlertPoliciesListByServerResponse,
   ServerSecurityAlertPoliciesListByServerNextResponse
 } from "../models";
 
@@ -64,11 +65,15 @@ export class ServerSecurityAlertPoliciesImpl
       [Symbol.asyncIterator]() {
         return this;
       },
-      byPage: () => {
+      byPage: (settings?: PageSettings) => {
+        if (settings?.maxPageSize) {
+          throw new Error("maxPageSize is not supported by this operation.");
+        }
         return this.listByServerPagingPage(
           resourceGroupName,
           serverName,
-          options
+          options,
+          settings
         );
       }
     };
@@ -77,15 +82,18 @@ export class ServerSecurityAlertPoliciesImpl
   private async *listByServerPagingPage(
     resourceGroupName: string,
     serverName: string,
-    options?: ServerSecurityAlertPoliciesListByServerOptionalParams
+    options?: ServerSecurityAlertPoliciesListByServerOptionalParams,
+    settings?: PageSettings
   ): AsyncIterableIterator<ServerSecurityAlertPolicy[]> {
-    let result = await this._listByServer(
-      resourceGroupName,
-      serverName,
-      options
-    );
-    yield result.value || [];
-    let continuationToken = result.nextLink;
+    let result: ServerSecurityAlertPoliciesListByServerResponse;
+    let continuationToken = settings?.continuationToken;
+    if (!continuationToken) {
+      result = await this._listByServer(resourceGroupName, serverName, options);
+      let page = result.value || [];
+      continuationToken = result.nextLink;
+      setContinuationToken(page, continuationToken);
+      yield page;
+    }
     while (continuationToken) {
       result = await this._listByServerNext(
         resourceGroupName,
@@ -94,7 +102,9 @@ export class ServerSecurityAlertPoliciesImpl
         options
       );
       continuationToken = result.nextLink;
-      yield result.value || [];
+      let page = result.value || [];
+      setContinuationToken(page, continuationToken);
+      yield page;
     }
   }