+++ sdk/communication/communication-phone-numbers/CHANGELOG.md
@@ -4,12 +4,14 @@
 
 ### Features Added
 
-- Updated to `@azure/core-tracing` 1.0.
-
 ### Breaking Changes
 
 ### Bugs Fixed
 
+### Other Changes
+
+- Updated to `@azure/core-tracing` 1.0.
+- Updated to `@azure/communication-common` 2.2.0.
 
 ## 1.2.0-beta.4 (Unreleased)